- Merging the changes to the master branch to the front end branch - Modified activity to fragment
