--- frontend/app/src/main/java/com/example/frontend/MapsActivity.kt
@@ -1,243 +1,0 @@
-package com.example.frontend
-
-import android.location.Location
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.ActivityCompat.OnRequestPermissionsResultCallback
-import androidx.core.content.ContextCompat
-import androidx.lifecycle.lifecycleScope
-import com.example.frontend.databinding.ActivityMapsBinding
-import com.example.frontend.place.Place
-import com.example.frontend.place.PlaceRenderer
-import com.example.frontend.place.PlacesReader
-import com.google.android.gms.maps.CameraUpdateFactory
-import com.google.android.gms.maps.GoogleMap
-import com.google.android.gms.maps.GoogleMap.OnMyLocationButtonClickListener
-import com.google.android.gms.maps.GoogleMap.OnMyLocationClickListener
-import com.google.android.gms.maps.OnMapReadyCallback
-import com.google.android.gms.maps.SupportMapFragment
-import com.google.android.gms.maps.model.BitmapDescriptor
-import com.google.android.gms.maps.model.LatLngBounds
-import com.google.maps.android.clustering.ClusterManager
-import com.google.maps.android.ktx.addMarker
-import com.google.maps.android.ktx.awaitMap
-import com.google.maps.android.ktx.awaitMapLoad
-import android.Manifest
-import android.annotation.SuppressLint
-import android.content.pm.PackageManager
-import android.widget.Toast
-import androidx.core.app.ActivityCompat
-import com.example.frontend.utils.PermissionUtils
-import com.example.frontend.utils.PermissionUtils.PermissionDeniedDialog.Companion.newInstance
-import com.example.frontend.utils.PermissionUtils.isPermissionGranted
-
-
-class MapsActivity : AppCompatActivity(),
-    OnMapReadyCallback,
-    OnMyLocationButtonClickListener,
-    OnMyLocationClickListener,
-    OnRequestPermissionsResultCallback
-{
-
-    private lateinit var mMap: GoogleMap
-    private val binding by lazy { ActivityMapsBinding.inflate(layoutInflater) }
-    private var permissionDenied = false
-
-    private val places: List<Place> by lazy {
-        PlacesReader(this).read()
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        val mapFragment = supportFragmentManager.findFragmentById(R.id.map) as SupportMapFragment
-        lifecycleScope.launchWhenCreated {
-            // Get map
-            val googleMap = mapFragment.awaitMap()
-
-            // Wait for map to finish loading
-            googleMap.awaitMapLoad()
-
-            // Ensure all places are visible in the map
-            val bounds = LatLngBounds.builder()
-            places.forEach { bounds.include(it.latLng) }
-            googleMap.moveCamera(CameraUpdateFactory.newLatLngBounds(bounds.build(), 20))
-
-            addClusteredMarkers(googleMap)
-            googleMap.setInfoWindowAdapter(MarkerInfoWindowAdapter(this@MapsActivity))
-        }
-
-    }
-
-    override fun onMapReady(googleMap: GoogleMap) {
-        mMap = googleMap
-        enableMyLocation()
-    }
-
-    private val trashIcon: BitmapDescriptor by lazy {
-        val color = ContextCompat.getColor(this, R.color.primary)
-        BitmapHelper.vectorToBitmap(this, R.drawable.ic_baseline_restore_from_trash_24, color)
-    }
-
-
-    /**
-     * Adds markers to the map. These markers won't be clustered.
-     */
-    private fun addMarkers(googleMap: GoogleMap) {
-        places.forEach { place ->
-            val marker = googleMap.addMarker {
-                title(place.name)
-                    .position(place.latLng)
-                    .icon(trashIcon)
-            }
-            // Set place as the tag on the marker object so it can be referenced within
-            // MarkerInfoWindowAdapter
-            marker.tag = place
-        }
-
-    }
-
-    /**
-     * Adds markers to the map with clustering support.
-     */
-    private fun addClusteredMarkers(googleMap: GoogleMap) {
-        // Create the ClusterManager class and set the custom renderer.
-        val clusterManager = ClusterManager<Place>(this, googleMap)
-        clusterManager.renderer =
-            PlaceRenderer(
-                this,
-                googleMap,
-                clusterManager
-            )
-
-        // Set custom info window adapter
-        clusterManager.markerCollection.setInfoWindowAdapter(MarkerInfoWindowAdapter(this))
-
-        // Add the places to the ClusterManager.
-        clusterManager.addItems(places)
-        clusterManager.cluster()
-
-        // When the camera starts moving, change the alpha value of the marker to translucent
-        googleMap.setOnCameraMoveStartedListener {
-            clusterManager.markerCollection.markers.forEach { it.alpha = 0.3f }
-            clusterManager.clusterMarkerCollection.markers.forEach { it.alpha = 0.3f }
-        }
-
-        // Set ClusterManager as the OnCameraIdleListener so that it
-        // can re-cluster when zooming in and out.
-        googleMap.setOnCameraIdleListener {
-            // When the camera stops moving, change the alpha value back to opaque
-            clusterManager.markerCollection.markers.forEach { it.alpha = 1.0f }
-            clusterManager.clusterMarkerCollection.markers.forEach { it.alpha = 1.0f }
-
-            clusterManager.onCameraIdle()
-
-        }
-    }
-
-    @SuppressLint("MissingPermission")
-    private fun enableMyLocation(){
-        // 1. Permissions are granted -> enable location
-        if(ContextCompat.checkSelfPermission(
-                this,
-                Manifest.permission.ACCESS_FINE_LOCATION
-            ) == PackageManager.PERMISSION_GRANTED || ContextCompat.checkSelfPermission(
-                this,
-                Manifest.permission.ACCESS_COARSE_LOCATION
-            ) == PackageManager.PERMISSION_GRANTED
-        ){
-            mMap.isMyLocationEnabled = true
-            return
-        }
-
-        // 2.
-        if (ActivityCompat.shouldShowRequestPermissionRationale(
-                this,
-                Manifest.permission.ACCESS_FINE_LOCATION
-            ) || ActivityCompat.shouldShowRequestPermissionRationale(
-                this,
-                Manifest.permission.ACCESS_COARSE_LOCATION
-            )
-        ) {
-            PermissionUtils.RationaleDialog.newInstance(
-                LOCATION_PERMISSION_REQUEST_CODE, true
-            ).show(supportFragmentManager, "dialog")
-            return
-        }
-
-        // 3. Otherwise, request permission
-        ActivityCompat.requestPermissions(
-            this,
-            arrayOf(
-                Manifest.permission.ACCESS_FINE_LOCATION,
-                Manifest.permission.ACCESS_COARSE_LOCATION
-            ),
-            LOCATION_PERMISSION_REQUEST_CODE
-        )
-
-
-    }
-
-    override fun onMyLocationButtonClick(): Boolean {
-        Toast.makeText(this, "MyLocation button clicked", Toast.LENGTH_SHORT)
-            .show()
-        // Return false so that we don't consume the event and the default behavior still occurs
-        // (the camera animates to the user's current position).
-        return false
-    }
-
-    override fun onMyLocationClick(location: Location) {
-        Toast.makeText(this, "Current location:\n$location", Toast.LENGTH_LONG)
-            .show()
-    }
-
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<String>,
-        grantResults: IntArray
-    ) {
-        if (requestCode != LOCATION_PERMISSION_REQUEST_CODE){
-            super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-            return
-        }
-
-        if (isPermissionGranted(
-                permissions, grantResults,Manifest.permission.ACCESS_FINE_LOCATION
-        ) || isPermissionGranted(
-                permissions, grantResults, Manifest.permission.ACCESS_COARSE_LOCATION
-            )
-        ) {
-            // Enable the my location layer if the permission has been granted.
-            enableMyLocation()
-        } else {
-            // Permission was denied. Display an error message
-            // Display the missing permission error dialog when the fragments resume.
-            permissionDenied = true
-        }
-    }
-
-    override fun onResumeFragments() {
-        super.onResumeFragments()
-        if (permissionDenied){
-            showMissingPermissionError()
-            permissionDenied = false
-        }
-    }
-
-    /**
-     * Displays a dialog with error message explaining that the location permission is missing.
-     */
-    private fun showMissingPermissionError() {
-        newInstance(true).show(supportFragmentManager, "dialog")
-    }
-
-    companion object {
-        /**
-         * Request code for location permission request.
-         *
-         * @see .onRequestPermissionsResult
-         */
-        private const val LOCATION_PERMISSION_REQUEST_CODE = 1
-    }
-
-}
