+++ frontend/app/src/main/res/values/colors.xml
@@ -7,6 +7,5 @@
     <color name="teal_200">#FF03DAC5</color>
     <color name="teal_700">#FF018786</color>
     <color name="black">#222B45</color>
-    <color name="primary">#EF1D52</color>
     <color name="white">#FFFFFFFF</color>
 </resources>
